backend_python/requirements.txt
Flask==2.3.0
flask-cors==5.0.1
mysql-connector-python==8.0.33
python-dotenv==1.0.0
werkzeug==2.3.0
cloudinary==1.36.0
requests==2.31.0
pyjwt==2.8.0
fpdf
